+++ main.php
@@ -2,6 +2,7 @@
 
 // ファイルのロード
 require_once('./lib/Loader.php');
+require_once('./lib/Utility.php');
 
 // オートロード
 $loader = new Loader();
@@ -24,22 +25,6 @@ $turn = 1;
 $isFinishFlg = false;
 
 $messageObj = new Message;
-
-function isFinish($objects)
-{
-  $deathCnt = 0; // HPが0以下の仲間の数
-  foreach($objects as $object){
-    // 1人でもHPが0を超えていたらfalseを返す
-    if($object->getHitPoint() > 0){
-      return false;
-    }
-    $deathCnt++;
-  }
-  // 仲間の数が死亡数(HPが0以下の数)と同じであればtrueを返す
-  if($deathCnt === count($objects)){
-    return true;
-  }
-}
 
 // どちらかのHPが0になるまで繰り返す
 while(!$isFinishFlg)
